requirements.txt
setuptools
wheel
flask
pyyaml
smbus2